pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gitops-drift"
version = "0.1.0"
description = "GitOps drift detection controller for Kubernetes"
requires-python = ">=3.9"
dependencies = [
    "kubernetes>=28.1.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.scripts]
gitops-drift = "gitops_drift.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
